{"schema":"decodex/commit/1","summary":"Split quantitative product row checks","authority":"manual"}
